introduce virt-config.go
